Fixed Compile error and Update common_utils

+++ include/pifod_ros2/imu_posture_ekf.hpp
@@ -16,36 +16,36 @@ namespace pifod_ros2
         /// @param angular 角速度[rad]
         /// @param linear_accel 加速度[m/s^2]
         /// @return オイラー角(roll, pitch, yaw)
-        Eigen::Vector3d estimate(Eigen::Vector3d angular, Eigen::Vector3d linear_accel);
+        Eigen::Vector3f estimate(Eigen::Vector3f angular, Eigen::Vector3f linear_accel);
 
         private:
-        Eigen::Matrix3d cov_;
-        Eigen::Matrix3d estimation_noise_;
-        Eigen::Matrix2d observation_noise_;
-        Eigen::MatrixXd kalman_gain_;
-        Eigen::Vector3d estimation_;
+        Eigen::Matrix3f cov_;
+        Eigen::Matrix3f estimation_noise_;
+        Eigen::Matrix2f observation_noise_;
+        Eigen::MatrixXf kalman_gain_;
+        Eigen::Vector3f estimation_;
     };
-    Eigen::Vector3d getInputMatrix(const double &angular_x, const double &angular_y, const double &angular_z);
+    Eigen::Vector3f getInputMatrix(const float &angular_x, const float &angular_y, const float &angular_z);
 
-    Eigen::Matrix<double, 3, 2> h();
+    Eigen::Matrix<float, 3, 2> h();
 
-    Eigen::Matrix3d jacob(const Eigen::Vector3d &input_matrix, const Eigen::Vector3d &estimation);
+    Eigen::Matrix3f jacob(const Eigen::Vector3f &input_matrix, const Eigen::Vector3f &estimation);
 
-    Eigen::Vector3d predictX(const Eigen::Vector3d &input_matrix, const Eigen::Vector3d &estimation);
+    Eigen::Vector3f predictX(const Eigen::Vector3f &input_matrix, const Eigen::Vector3f &estimation);
 
-    Eigen::Matrix3d predictCov(const Eigen::Matrix3d &jacob, const Eigen::Matrix3d &cov, const Eigen::Matrix3d &est_noise);
+    Eigen::Matrix3f predictCov(const Eigen::Matrix3f &jacob, const Eigen::Matrix3f &cov, const Eigen::Matrix3f &est_noise);
 
-    Eigen::Vector2d updateResidual(const Eigen::Vector2d &obs, const Eigen::Vector3d &est);
+    Eigen::Vector2f updateResidual(const Eigen::Vector2f &obs, const Eigen::Vector3f &est);
 
-    Eigen::Matrix2d updateS(const Eigen::Matrix3d &cov_, const Eigen::Matrix2d &obs_noise);
+    Eigen::Matrix2f updateS(const Eigen::Matrix3f &cov_, const Eigen::Matrix2f &obs_noise);
 
-    Eigen::Matrix<double, 3, 2> updateKalmanGain(const Eigen::Matrix2d &s, const Eigen::Matrix3d &cov);
+    Eigen::Matrix<float, 3, 2> updateKalmanGain(const Eigen::Matrix2f &s, const Eigen::Matrix3f &cov);
 
-    Eigen::Vector3d updateX(const Eigen::Vector3d &est, const Eigen::Matrix<double, 3, 2> &kalman_gain_, const Eigen::Vector2d &residual);
+    Eigen::Vector3f updateX(const Eigen::Vector3f &est, const Eigen::Matrix<float, 3, 2> &kalman_gain_, const Eigen::Vector2f &residual);
 
-    Eigen::Matrix3d updateCov(const Eigen::Matrix<double, 3, 2> &kalman_gain, const Eigen::Matrix3d &cov);
+    Eigen::Matrix3f updateCov(const Eigen::Matrix<float, 3, 2> &kalman_gain, const Eigen::Matrix3f &cov);
 
-    Eigen::Vector2d obsModel(const Eigen::Vector3d &linear_accel);
+    Eigen::Vector2f obsModel(const Eigen::Vector3f &linear_accel);
 }
 
 #endif
